+++ run_tests.sh
@@ -1,4 +1,4 @@
 #! /usr/bin/env sh
 export PYTHONPATH="src:$PYTHONPATH"
 
-py.test tests $@
+py.test3 tests $@