+++ src/dhtmlparser/htmlelement/__init__.py
@@ -4,8 +4,8 @@
 # Interpreter version: python 2.7
 #
 # Imports =====================================================================
-from shared import NONPAIR_TAGS
-from shared import _rotate_buff
-from shared import _closeElements
+from .shared import NONPAIR_TAGS
+from .shared import _rotate_buff
+from .shared import _closeElements
 
-from html_element import HTMLElement
+from .html_element import HTMLElement