+++ src/dhtmlparser/htmlelement/html_parser.py
@@ -10,15 +10,18 @@ HTMLElement class used in DOM representation.
 from ..quoter import escape, unescape
 from ..specialdict import SpecialDict
 
-from shared import NONPAIR_TAGS
-from shared import _rotate_buff
-from shared import _closeElements
+from .shared import NONPAIR_TAGS
+from .shared import _rotate_buff
+from .shared import _closeElements
 
 
 # Functions & objects =========================================================
 # helper functions
 def _is_str(tag):
-    return isinstance(tag, basestring)
+    try:
+        return isinstance(tag, basestring)
+    except NameError:
+        return isinstance(tag, str)
 
 
 def _is_dict(tag):