+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.1.8
 -----
     - Fixed python 2 / 3 problem in docs/__init__.py.
+    - getVersion() -> get_version().
 
 2.1.7
 -----